--- Include/AxonEngine.h
@@ -1,11 +1,0 @@
-﻿#pragma once
-
-#if defined(AXON_SERVER)
-#include <AxonServer.h>
-#elif defined(AXON_CLIENT)
-#include <AxonClient.h>
-#endif
-
-#include <Axon.h>
-#include <AxonError.h>
-#include <AxonTypes.h>
