--- Include/AxonError.cpp
@@ -1,6 +1,0 @@
-#include "AxonError.h"
-
-
-Axon::AxonError::AxonError(Axon::Error::AxonErrorCode err) : reason(err) {}
-
-constexpr Axon::Error::AxonErrorCode Axon::AxonError::err() const { return reason; }
