--- build-scripts/server/post-build.bat
@@ -1,1 +1,0 @@
-set WORKING_DIR="./out/"
