--- examples/02. basic-gui/server.cpp
@@ -1,11 +1,0 @@
-#ifndef AXON_SERVER
-#define AXON_SERVER
-#endif
-
-#include <AxonEngine.h>
-
-
-int main()
-{
-	Axon::Server server = Axon::Server();
-}
