--- examples/README.md
@@ -1,17 +1,0 @@
-# Examples
-
-## Install
-
-First of all, build your Axon Engine. You should get .dll/.lib for Windows target platform and .a/.so for UNIX systems. 
-Modify CMakeLists.txt if you have your build stored in custom directory.
-
-## Building
-
-Building example:
-
-```bash
-	mkdir cmake-build
-	cd cmake-build
-
-	cmake .. && cmake --build .
-```
