--- tests/docker-test/compose.yaml
@@ -1,15 +1,0 @@
-version: '3.7'
-
-services:
-    axon-server:
-        build: 
-            context: ../../
-            dockerfile: ./tests/docker-test/server/Dockerfile
-        container_name: server
-    axon-client:
-        build:
-            context: ../../
-            dockerfile: ./tests/docker-test/client/Dockerfile
-        depends_on:
-            - axon-server
-        container_name: client
