--- tests/docker-test/server/Dockerfile
@@ -1,10 +1,0 @@
-FROM ubuntu:latest
-
-COPY cmake-build-debug/AxonDockertestServer /home/ubuntu
-COPY cmake-build-debug/libAxonServerLib.so /usr/local/lib
-
-RUN ldconfig
-
-ENTRYPOINT [ "./home/ubuntu/AxonDockertestServer" ]
-
-CMD ["run"]
